# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Banking Eligibility Assistant API
  description: |
    Mock APIs و موتور تطبیق اهلیت محصولات بانکی.
    شامل سرویس‌های هویتی، مالی، RBCI، محصولات، بخشنامه‌ها و matching engine.
  version: 1.0.0
  contact:
    name: DATA-ICTChallenge
servers:
  - url: http://localhost:8080
    description: Local
  - url: http://nl-main.z3df1lter.uk:9080
    description: Production VPS
paths:
...
  /api/rbci/cold-start:
    post:
      summary: ارزیابی ریسک اولیه (غیرمشتری)
      operationId: coldStartRisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColdStartRequest'
      responses:
        '200':
          description: ارزیابی ریسک اولیه
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAssessment'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/circulars/by-product:
    get:
      summary: بخشنامه‌های یک محصول
      operationId: circularsByProduct
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: string
            example: "P001"
      responses:
        '200':
          description: قواعد محصول
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CircularRule'
...
components:
  responses:
    BadRequest:
      description: پارامتر نامعتبر
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        error_fa:
          type: string
...
    RiskAssessment:
      type: object
      properties:
        customer_id: { type: string }
        risk_level: { type: string, enum: [low, medium, high] }
        risk_score: { type: number }
        reason: { type: string }
        is_cold_start: { type: boolean }
    ColdStartRequest:
      type: object
      required: [age, occupation]
      properties:
        name: { type: string }
        age: { type: integer }
        gender: { type: string }
        occupation: { type: string }
        employment_type: { type: string }
        approx_income: { type: number }
        visit_purpose: { type: string }
...
    CircularRule:
      type: object
      properties:
        id: { type: string }
        circular_ref: { type: string }
        circular_ref_fa: { type: string }
        product_id: { type: string }
        conditions:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              operator: { type: string }
              value: {}
        description: { type: string }
        description_fa: { type: string }
...